2010-11-09 10:47 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/src/rdd/sdf1.c     ! fixed bug in SDF import procedure - many thanks       to Enrico Maria Giordano for reporting the problem       and to Vicente Guerra for locating the reason       BTW fix committed to xHarbour may work but it's not correct,       I suggest to update it.
  * harbour/src/rtl/hbdyn.c
  * harbour/src/rtl/hbsocket.c
  * harbour/src/rdd/dbfnsx/dbfnsx1.c
  * harbour/src/rdd/dbfcdx/dbfcdx1.c
  * harbour/src/rdd/dbffpt/dbffpt1.c
  * harbour/contrib/hbct/bitnum.c
  * harbour/contrib/hbwin/olecore.c
    * pacified some of MSVC6 warnings reported by Andi

  * harbour/utils/Makefile
    * added internal/developers build switch

+++ harbour/src/rtl/hbsocket.c
@@ -1436,7 +1436,7 @@ static int hb_socketSelectRD( HB_SOCKET sd, HB_MAXINT timeout )
       if( timeout >= 0 )
       {
          tv.tv_sec = ( long ) ( timeout / 1000 );
-         tv.tv_usec = ( timeout % 1000 ) * 1000;
+         tv.tv_usec = ( long ) ( timeout % 1000 ) * 1000;
          ptv = &tv;
       }
       else
@@ -1484,7 +1484,7 @@ static int hb_socketSelectWR( HB_SOCKET sd, HB_MAXINT timeout )
       if( timeout >= 0 )
       {
          tv.tv_sec = ( long ) ( timeout / 1000 );
-         tv.tv_usec = ( timeout % 1000 ) * 1000;
+         tv.tv_usec = ( long ) ( timeout % 1000 ) * 1000;
          ptv = &tv;
       }
       else
@@ -1541,7 +1541,7 @@ static int hb_socketSelectWRE( HB_SOCKET sd, HB_MAXINT timeout )
       if( timeout >= 0 )
       {
          tv.tv_sec = ( long ) ( timeout / 1000 );
-         tv.tv_usec = ( timeout % 1000 ) * 1000;
+         tv.tv_usec = ( long ) ( timeout % 1000 ) * 1000;
          ptv = &tv;
       }
       else
@@ -2628,7 +2628,7 @@ int hb_socketSelect( PHB_ITEM pArrayRD, HB_BOOL fSetRD,
    if( timeout >= 0 )
    {
       tv.tv_sec = ( long ) ( timeout / 1000 );
-      tv.tv_usec = ( timeout % 1000 ) * 1000;
+      tv.tv_usec = ( long ) ( timeout % 1000 ) * 1000;
       ptv = &tv;
    }
    else
